moved format logic to seperate file
