+++ packages/cli/cmd/application/main.go
@@ -16,7 +16,6 @@ import (
 	"github.com/aws/amazon-genomics-cli/internal/pkg/cli"
 	"github.com/aws/amazon-genomics-cli/internal/pkg/cli/clierror"
 	"github.com/aws/amazon-genomics-cli/internal/pkg/cli/config"
-	"github.com/aws/amazon-genomics-cli/internal/pkg/cli/format"
 	"github.com/aws/amazon-genomics-cli/internal/pkg/logging"
 	"github.com/aws/amazon-genomics-cli/internal/pkg/storage"
 	"github.com/aws/amazon-genomics-cli/internal/pkg/term/color"
@@ -134,35 +133,6 @@ func buildRootCmd() *cobra.Command {
 	cmd.Flag("docs").Hidden = true
 
 	return cmd
-}
-
-func ValidateFormat(f format.FormatterType) error {
-	if err := f.ValidateFormatter(); err != nil {
-		return err
-	}
-	return nil
-}
-
-func setFormatter(f formatVars) string {
-	configClient, err := newConfigClient()
-	if err != nil {
-		log.Error().Err(err)
-		return ""
-	}
-	if f.format == "" {
-		f.format = defaultFormat
-		configFormat, err := configClient.GetFormat()
-		if err != nil {
-			log.Error().Err(err)
-		} else {
-			f.format = configFormat
-		}
-	}
-	if err := ValidateFormat(format.FormatterType(f.format)); err != nil {
-		fmt.Println(err.Error())
-	}
-	format.SetFormatter(format.FormatterType(f.format))
-	return f.format
 }
 
 func checkCliVersion() {
